# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ioc-cfn-mgmt-backend API",
    "description": "IOC CFN Management Backend Service API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/workspaces/{workspace_id}/invitations": {
      "post": {
        "tags": [
          "workspace-invitations"
        ],
        "summary": "Create Invitation",
        "description": "Create a workspace invitation (admin only)\n\n- **workspace_id**: UUID of the workspace\n- **invitee_username**: Username of user to invite\n- **role**: Role to assign (admin/viewer/guest)\n\nReturns the UUID of the created invitation",
        "operationId": "create_invitation_api_workspaces__workspace_id__invitations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceInvitationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInvitationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workspace-invitations"
        ],
        "summary": "List Workspace Invitations",
        "description": "List all invitations for a workspace (admin only)\n\n- **workspace_id**: UUID of the workspace\n\nReturns a list of all invitations (pending, accepted, declined, expired)",
        "operationId": "list_workspace_invitations_api_workspaces__workspace_id__invitations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInvitationList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/invitations/{invitation_id}": {
      "delete": {
        "tags": [
          "workspace-invitations"
        ],
        "summary": "Cancel Invitation",
        "description": "Cancel a pending invitation (admin only)\n\n- **workspace_id**: UUID of the workspace\n- **invitation_id**: UUID of the invitation\n\nReturns success message",
        "operationId": "cancel_invitation_api_workspaces__workspace_id__invitations__invitation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invitation Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invitations/pending": {
      "get": {
        "tags": [
          "workspace-invitations"
        ],
        "summary": "Get Pending Invitations",
        "description": "Get all pending invitations for the current user\n\nReturns a list of pending invitations",
        "operationId": "get_pending_invitations_api_invitations_pending_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInvitationList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invitations/{invitation_id}/accept": {
      "post": {
        "tags": [
          "workspace-invitations"
        ],
        "summary": "Accept Invitation",
        "description": "Accept a workspace invitation\n\n- **invitation_id**: UUID of the invitation\n\nReturns success message with workspace details and assigned role",
        "operationId": "accept_invitation_api_invitations__invitation_id__accept_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invitation Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInvitationAcceptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invitations/{invitation_id}/decline": {
      "post": {
        "tags": [
          "workspace-invitations"
        ],
        "summary": "Decline Invitation",
        "description": "Decline a workspace invitation\n\n- **invitation_id**: UUID of the invitation\n\nReturns success message",
        "operationId": "decline_invitation_api_invitations__invitation_id__decline_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Invitation Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvitationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "accepted",
          "declined",
          "expired"
        ],
        "title": "InvitationStatus",
        "description": "Invitation status values"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkspaceInvitationAcceptResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "workspace_name": {
            "type": "string",
            "title": "Workspace Name"
          },
          "assigned_role": {
            "type": "string",
            "title": "Assigned Role"
          }
        },
        "type": "object",
        "required": [
          "message",
          "workspace_id",
          "workspace_name",
          "assigned_role"
        ],
        "title": "WorkspaceInvitationAcceptResponse",
        "description": "Schema for invitation acceptance response"
      },
      "WorkspaceInvitationCreate": {
        "properties": {
          "invitee_username": {
            "type": "string",
            "title": "Invitee Username",
            "description": "Username of user to invite"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Role to assign (admin/viewer/guest)"
          }
        },
        "type": "object",
        "required": [
          "invitee_username",
          "role"
        ],
        "title": "WorkspaceInvitationCreate",
        "description": "Schema for creating a workspace invitation"
      },
      "WorkspaceInvitationDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "workspace_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Name",
            "description": "Workspace name joined from workspace table"
          },
          "inviter_id": {
            "type": "string",
            "title": "Inviter Id"
          },
          "inviter_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inviter Username",
            "description": "Inviter username joined from user table"
          },
          "invitee_username": {
            "type": "string",
            "title": "Invitee Username"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "status": {
            "$ref": "#/components/schemas/InvitationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "responded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "inviter_id",
          "invitee_username",
          "role",
          "status",
          "created_at",
          "expires_at"
        ],
        "title": "WorkspaceInvitationDetail",
        "description": "Schema for workspace invitation details"
      },
      "WorkspaceInvitationList": {
        "properties": {
          "invitations": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceInvitationDetail"
            },
            "type": "array",
            "title": "Invitations"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "invitations",
          "total"
        ],
        "title": "WorkspaceInvitationList",
        "description": "Schema for list of workspace invitations"
      },
      "WorkspaceInvitationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "WorkspaceInvitationResponse",
        "description": "Schema for invitation creation response"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}